# swagger/swagger.yml
openapi: 3.0.1
info:
  title: Yo!Photo 👻
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
  - url: http://localhost:8000/api/v1
  - url: http://localhost:8000/api/v2
...
paths:
...
  /albums:
    post:
      tags:
        - Album
      summary: Create an album
      requestBody:
        description: Create an album
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Swimming in Sabana, Merauke
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Swimming in Sabana, Merauke
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Album"
    get:
      tags:
        - Album
      summary: Get all albums in the database, with skip and take query, ordered by createdAt
      description: |
        Get all albums in the database, with skip and take query, ordered by createdAt
      parameters:
        - name: skip
          in: query
          description: skip
          required: true
          schema:
            type: integer
            example: 0
        - name: take
          in: query
          description: take
          required: true
          schema:
            type: integer
            example: 5
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Album"
    put:
      tags:
        - Album
      summary: Update detail of specific album, by album Id
      requestBody:
        description: Update detail of specific album, by album Id
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumDetail"
            example:
              albumId: f503adb6-3bd4-48a2-9b24-77eb1bae8238
              title: Flying in The Dark
              description: This is how I fly in the Dark
              coverPhotoId: 7
              isPrivate: true
              isDownloadable: true
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Album"
...
  /contest/community:
    get:
      tags:
        - CommunityContest
      summary: Get many with skip & take
      parameters:
        - name: isActive
          in: query
          description: active / inactive contest
          required: false
          schema:
            type: boolean
            example: true
        - name: skip
          in: query
          description: skip
          required: false
          schema:
            type: integer
            example: 0
        - name: take
          in: query
          description: take
          required: false
          schema:
            type: integer
            example: 5
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                # $ref: "#/components/schemas/Album"
                # $ref: "#/components/examples/Coord"

    post:
      tags:
        - CommunityContest
      summary: Create Community Contest
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Landscape Bekasi
                photo:
                  type: array
                  items:
                    type: string
                    format: binary
                constestCategoryId:
                  type: string
                  enum: [macro6i2tnd9, streethg5nnbz, nightp0annjf, landscapeuc0nnnd]
                  example: streethg5nnbz
                startAt:
                  type: string
                  format: date-time
                  example: 2022-10-06T23:16:15.850Z
                endAt:
                  type: string
                  format: date-time
                  example: 2022-10-06T23:16:15.850Z
                ratingStartAt:
                  type: string
                  format: date-time
                  example: 2022-10-06T23:16:15.850Z
                ratingEndtAt:
                  type: string
                  format: date-time
                  example: 2022-10-06T23:16:15.850Z
                isJudged:
                  type: boolean
                  example: false
                brief:
                  type: string
                  example: A diary writing is a personal record of events and feelings 
                    affecting the writer.\n It is written for personal record and reflection
                    therefore, it may not always be objective in its view. It almost 
                    always details the events occurring in the present. 
                    It includes facts, observations, perceptions, and descriptions all 
                    according to the writer’s worldview.
                tags:
                  type: string
                  example: Louis,Vuitton
                isActive:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                # $ref: "#/components/schemas/Album"
                # $ref: "#/components/examples/Coord"
components:
...
  schemas:
...
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: a40568c5-174f-4b34-8f9c-e0dedeab0cd8
        title:
          type: string
          example: Wrapsafe
        description:
          type: string
          example: Inverse encompassing extranet
        coverPhotoId:
          type: string
          format: uuid
          example: 00771301-989b-4599-aed0-b1a64dc2d23c
        isPrivate:
          type: boolean
          example: false
        isDownloadable:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: 2021-05-11T03:44:35.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2021-05-11T05:56:10.000Z
...
    UpdateAlbumDetail:
      type: object
      properties:
        albumId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        coverPhotoId:
          type: string
        isPrivate:
          type: boolean
        isDownloadable:
          type: boolean
...